fpga

+++ src/caffe/layers/data_layer.cpp
@@ -217,8 +217,8 @@ DataLayer<Ftype, Btype>::DataLayerSetUp(const vector<Blob*>& bottom, const vecto
   // newplan added
   if(this->phase_ == TRAIN)
   {
-    const size_t cropped_height = param.transform_param().crop_size();
-    const size_t cropped_width = param.transform_param().crop_size();
+    const int cropped_height = param.transform_param().crop_size();
+    const int cropped_width = param.transform_param().crop_size();
     //Packing packing = NHWC;  // OpenCV
     vector<int> top_shape = {batch_size,new_channel,cropped_height,cropped_width};
     top[0]->Reshape(top_shape);
